+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - updated test cases in search options
 - updated search profile save rating function to use a string instead of an int
 - filter function working in that it pulls back a list of movies
+- added getMovies, getYears, and getRatings to the metaData class
 
 ### 4/7/23
 - Uploaded gui.py